+++ spec/lite/command/base_spec.rb
@@ -250,9 +250,9 @@ RSpec.describe Lite::Command::Base do
 
       it "returns a noop status" do
         expect(command).to be_noop
-        expect(command).to be_faulter
-        expect(command).to be_thrower
-        expect(command).not_to be_thrown_fault
+        expect(command).to be_origin
+        expect(command).to be_source
+        expect(command).not_to be_thrown
         expect(command.noop?("[!] command stopped due to noop")).to be(true)
         expect(command.noop?("Some reason")).to be(false)
         expect(command.status).to eq(Lite::Command::NOOP)
@@ -264,9 +264,9 @@ RSpec.describe Lite::Command::Base do
 
       it "returns a error status" do
         expect(command).to be_invalid
-        expect(command).to be_faulter
-        expect(command).to be_thrower
-        expect(command).not_to be_thrown_fault
+        expect(command).to be_origin
+        expect(command).to be_source
+        expect(command).not_to be_thrown
         expect(command.invalid?("[!] command stopped due to invalid")).to be(true)
         expect(command.invalid?("Some reason")).to be(false)
         expect(command.status).to eq(Lite::Command::INVALID)
@@ -278,9 +278,9 @@ RSpec.describe Lite::Command::Base do
 
       it "returns a failure status" do
         expect(command).to be_failure
-        expect(command).to be_faulter
-        expect(command).to be_thrower
-        expect(command).not_to be_thrown_fault
+        expect(command).to be_origin
+        expect(command).to be_source
+        expect(command).not_to be_thrown
         expect(command.failure?("[!] command stopped due to failure")).to be(true)
         expect(command.failure?("Some reason")).to be(false)
         expect(command.status).to eq(Lite::Command::FAILURE)
@@ -292,9 +292,9 @@ RSpec.describe Lite::Command::Base do
 
       it "returns a error status" do
         expect(command).to be_error
-        expect(command).to be_faulter
-        expect(command).to be_thrower
-        expect(command).not_to be_thrown_fault
+        expect(command).to be_origin
+        expect(command).to be_source
+        expect(command).not_to be_thrown
         expect(command.error?("[!] command stopped due to error")).to be(true)
         expect(command.error?("Some reason")).to be(false)
         expect(command.status).to eq(Lite::Command::ERROR)
@@ -306,9 +306,9 @@ RSpec.describe Lite::Command::Base do
 
       it "returns a error status" do
         expect(command).to be_error
-        expect(command).to be_faulter
-        expect(command).to be_thrower
-        expect(command).not_to be_thrown_fault
+        expect(command).to be_origin
+        expect(command).to be_source
+        expect(command).not_to be_thrown
         expect(command.error?("[RuntimeError] [!] command stopped due to exception")).to be(true)
         expect(command.error?("Some reason")).to be(false)
         expect(command.status).to eq(Lite::Command::ERROR)
@@ -320,13 +320,13 @@ RSpec.describe Lite::Command::Base do
 
       it "returns fault" do
         expect(command).to be_fault
-        expect(command).not_to be_faulter
-        expect(command).not_to be_thrower
-        expect(command).to be_thrown_fault
+        expect(command).not_to be_origin
+        expect(command).not_to be_source
+        expect(command).to be_thrown
         expect(command.fault?("[!] command stopped due to child noop")).to be(true)
         expect(command.fault?("Some reason")).to be(false)
-        expect(command.faulter?).to be(false)
-        expect(command.thrower?).to be(false)
+        expect(command.origin?).to be(false)
+        expect(command.source?).to be(false)
       end
     end
   end
@@ -375,8 +375,8 @@ RSpec.describe Lite::Command::Base do
           state: "dnf",
           status: "noop",
           reason: "[!] command stopped due to noop",
-          fault: 1,
-          throw: 1,
+          origin: 1,
+          source: 1,
           runtime: 0.0123
         )
       end
@@ -395,8 +395,8 @@ RSpec.describe Lite::Command::Base do
           state: "dnf",
           status: "invalid",
           reason: "[!] command stopped due to invalid",
-          fault: 1,
-          throw: 1,
+          origin: 1,
+          source: 1,
           runtime: 0.0123
         )
       end
@@ -415,8 +415,8 @@ RSpec.describe Lite::Command::Base do
           state: "dnf",
           status: "failure",
           reason: "[!] command stopped due to failure",
-          fault: 1,
-          throw: 1,
+          origin: 1,
+          source: 1,
           runtime: 0.0123
         )
       end
@@ -435,8 +435,8 @@ RSpec.describe Lite::Command::Base do
           state: "dnf",
           status: "error",
           reason: "[!] command stopped due to error",
-          fault: 1,
-          throw: 1,
+          origin: 1,
+          source: 1,
           runtime: 0.0123
         )
       end
@@ -455,8 +455,8 @@ RSpec.describe Lite::Command::Base do
           state: "dnf",
           status: "error",
           reason: "[RuntimeError] [!] command stopped due to exception",
-          fault: 1,
-          throw: 1,
+          origin: 1,
+          source: 1,
           runtime: 0.0123
         )
       end
@@ -475,8 +475,8 @@ RSpec.describe Lite::Command::Base do
           state: "dnf",
           status: "noop",
           reason: "[!] command stopped due to child noop",
-          fault: 5,
-          throw: 5,
+          origin: 5,
+          source: 5,
           runtime: 0.0123
         )
       end